♻️ 유저 정보 전달 로직 api 위치 변경
