+++ app/backend/src/app.module.ts
@@ -2,6 +2,7 @@ import { Module } from '@nestjs/common';
 import { CacheModule } from '@nestjs/cache-manager';
 import { AuthModule } from './auth/auth.module';
 import { AppController } from './app.controller';
+import { MemberModule } from './member/member.module';
 import * as redisStore from 'cache-manager-redis-store';
 
 @Module({
@@ -13,6 +14,7 @@ import * as redisStore from 'cache-manager-redis-store';
       port: process.env.REDIS_PORT,
     }),
     AuthModule,
+    MemberModule,
   ],
   controllers: [AppController],
   providers: [],
