+++ app/backend/src/member/member.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { MemberService } from './member.service';
+import { MemberController } from './member.controller';
+import { JwtModule } from '@nestjs/jwt';
+
+@Module({
+  imports: [JwtModule],
+  controllers: [MemberController],
+  providers: [MemberService],
+})
+export class MemberModule {}
